Refactor lambda layer system with clean naming conventions
- Replace fast_lambda_layer.py with enhanced lambda_layer.py
- Rename FastLambdaLayer class to LambdaLayer for clean naming
- Add hybrid sync/async mode with real LayerVersion ARNs
- Fix null ARN errors in Lambda function references
- Maintain all advanced features: change detection, pipeline builds, layer optimization
- Backup original implementation as lambda_layer_old.py
- Enable sync mode (USE_SYNC_MODE = True) to provide real ARNs for stack compatibility
